% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.R
\name{top}
\alias{top}
\title{<brief desc>}
\usage{
top(mat, n = 5)
}
\arguments{
\item{mat}{a matrix or data.frame}

\item{n}{number of rows/columns to display}
}
\value{
a Matrix with dimension nxn
}
\description{
full desc for \code{\link{top}}
}
\examples{
\dontrun{ x=matrix(rnorm(1000, ncol=10); top(4))}
}
\seealso{
\code{\link[utils]{head}}
}
